+++ CHANGELOG.md
@@ -49,3 +49,4 @@ All notable changes to the "my-racing-planner" project will be documented in thi
 
 - Improved code splitting a bit by lay loading pages
 - Make tables scroll to top when changing pages
+- Disabled text selection on images and button texts